Working on build process

+++ scripts/test_cover.sh
@@ -1,11 +1,20 @@
 #!/bin/bash
 
+if [ -z "$1" ]; then
+    echo "Usage: $0 <path-to-executable>"
+    exit 1
+fi
+
+
+EXECUTABLE_PATH="$1"
+
+
 ############################
 ### Delete previous data ###
 ############################
 
 echo "Removing previous build/coverage directory..."
-rm -rf build/coverage || { echo "Failed to remove build/coverage"; exit 1; }
+rm -rf "$EXECUTABLE_PATH/coverage" || { echo "Failed to remove $EXECUTABLE_PATH/coverage"; exit 1; }
 
 echo "Deleting .gcda files..."
 find . -name "*.gcda" -delete || { echo "Failed to delete .gcda files"; exit 1; }
@@ -16,9 +25,7 @@ find . -name "*.gcda" -delete || { echo "Failed to delete .gcda files"; exit 1; 
 ########################
 
 echo "Running sqlighter_tests..."
-cd build/bin      || { echo "Failed to change directory to build/bin"; exit 1; }
-./sqlighter_tests || { echo "Test execution failed"; exit 1; }
-cd ../..          || { echo "Failed to navigate back to the project root"; exit 1; }
+"$EXECUTABLE_PATH/sqlighter_tests" || { echo "Test execution failed"; exit 1; }
 
 
 ########################
@@ -26,8 +33,8 @@ cd ../..          || { echo "Failed to navigate back to the project root"; exit 
 ########################
 
 echo "Generating coverage info..."
-lcov --capture --directory build --output-file build/coverage.info                              || { echo "Failed to generate coverage info"; exit 1; }
-lcov --extract build/coverage.info '*/sqlighter/*' --output-file build/sqlighter_coverage.info  || { echo "Failed to filter coverage info"; exit 1; }
-genhtml build/sqlighter_coverage.info --output-directory build/coverage                         || { echo "Failed to generate HTML report"; exit 1; }
+lcov --capture --directory "$EXECUTABLE_PATH" --output-file "$EXECUTABLE_PATH/coverage.info"                              || { echo "Failed to generate coverage info"; exit 1; }
+lcov --extract "$EXECUTABLE_PATH/coverage.info" '*/sqlighter/*' --output-file "$EXECUTABLE_PATH/sqlighter_coverage.info"  || { echo "Failed to filter coverage info"; exit 1; }
+genhtml "$EXECUTABLE_PATH/sqlighter_coverage.info" --output-directory "$EXECUTABLE_PATH/coverage"                         || { echo "Failed to generate HTML report"; exit 1; }
 
 echo "Script completed successfully."
